auto refresh tmux

+++ scripts/sys-theme.sh
@@ -4,19 +4,31 @@ function sync_theme() {
   local alacritty_config_file=$HOME/.config/alacritty/alacritty.toml
   local alacritty_theme_file=$HOME/.config/alacritty/theme.toml
 
+  local prev=$(cat $HOME/.theme)
+
   if [[ $(defaults read -g AppleInterfaceStyle 2> /dev/null) == "Dark" ]]
   then
-    echo "dark" > $HOME/.theme
+    if [[ $prev != "dark" ]]
+    then
+      echo -e "dark" > $HOME/.theme
 
-    rm $alacritty_theme_file && \
-      ln -s $HOME/.config/alacritty/themes/themes/everforest_dark.toml $alacritty_theme_file
-    touch $alacritty_config_file
+      rm $alacritty_theme_file && \
+        ln -s $HOME/.config/alacritty/themes/themes/everforest_dark.toml $alacritty_theme_file
+      touch $alacritty_config_file
+
+      tmux source $HOME/.tmux.conf
+    fi
   else
-    echo "light" > $HOME/.theme
+    if [[ $prev != "light" ]]
+    then
+      echo -e "light" > $HOME/.theme
 
-    rm $alacritty_theme_file && \
-      ln -s $HOME/.config/alacritty/themes/themes/everforest_light.toml $alacritty_theme_file
-    touch $alacritty_config_file
+      rm $alacritty_theme_file && \
+        ln -s $HOME/.config/alacritty/themes/themes/everforest_light.toml $alacritty_theme_file
+      touch $alacritty_config_file
+
+      tmux source $HOME/.tmux.conf
+    fi
   fi
 }
